tech-studio-projects/financial-industry/portfolio-management-simulation-platform/java/core-system/src/main/java/com/system/algotrading/execution/OrderExecution.java
package com.system.algotrading.execution;

import org.slf4j.Logger;
import org.slf4j.LoggerFactory;

/**
 * The OrderExecution class handles the processing and execution of trade orders.
 * This class is responsible for managing the lifecycle of an order, including validation
 * and logging of execution results.
 */
public class OrderExecution {

    // Logger instance to log order execution details and errors
    private static final Logger logger = LoggerFactory.getLogger(OrderExecution.class);

    /**
     * Executes the specified order after validating its contents. Logs the order execution
     * process and handles exceptions that may occur during order execution.
     *
     * @param order A string representation of the order to execute.
     * @return boolean indicating success (true) or failure (false) of the execution.
     */
    public boolean executeOrder(String order) {
        try {
            // Log the initial state of the order to be executed
            logger.info("Executing order: {}", order);

            // Validate the order before processing; throw an error if invalid
            if (order == null || order.isEmpty()) {
                throw new IllegalArgumentException("Order is invalid");
            }

            // Placeholder for actual execution logic
            // Low-latency operations would go here if this were a real system

            // Log a successful order execution for tracking
            logger.debug("Order executed successfully");
            return true;

        } catch (Exception e) {
            // Log error details if order execution fails
            logger.error("Order execution failed", e);
            return false;
        }
    }
}
